Update CLI documentation and refactor pull/push commands
This commit includes changes to the CLI's README.md file, improving the instructions for creating a `shelve.config.ts` file and updating the command list. It also introduces a JSON schema for configuration validation.

The pull and push commands in the CLI have been refactored. The code has been simplified by moving some logic into utility functions in `env.ts`. This includes creating or merging .env files during a pull operation, as well as pushing variables during a push operation.

Additionally, new configuration options have been added to `shelve.config.ts`, including `pushMethod`, `pullMethod`, and `envFileName`. These are used in the updated pull and push operations.
