+++ packages/cli/src/utils/env.ts
@@ -1,5 +1,5 @@
 import fs from 'fs'
-import type { Env } from '@shelve/types'
+import type { Env, VariablesCreateInput } from '@shelve/types'
 import { cancel, spinner } from '@clack/prompts'
 import consola from 'consola'
 import { loadShelveConfig } from './config'
@@ -13,25 +13,28 @@ export async function isEnvFileExist(): boolean {
 }
 
 export async function mergeEnvFile(variables: Env = []): void {
+  s.start('Merging .env file')
   const { envFileName } = await loadShelveConfig()
   const envFile = await getEnvFile()
   envFile.push(...variables)
   const content = envFile.map((item) => `${ item.key }=${ item.value }`).join('\n')
   const finalString = `# Generated by Shelve CLI\n${content}`
   fs.writeFileSync(envFileName, finalString)
+  s.stop('Merging .env file')
   consola.success(`${ envFileName } file updated successfully`)
 }
 
-export async function createEnvFile(variables: Env = []): void {
-  const { pullMethod, envFileName } = await loadShelveConfig()
+export async function createEnvFile(pullMethod: string, envFileName: string, variables: Env = []): Promise<void> {
   if (isEnvFileExist() && pullMethod === 'merge') {
     await mergeEnvFile(variables)
     return
   }
+  s.start('Creating .env file')
   const content = variables.map((item) => `${ item.key }=${ item.value }`).join('\n')
   const finalString = `# Generated by Shelve CLI\n${content}`
   if (isEnvFileExist()) fs.unlinkSync(envFileName)
   fs.writeFileSync(envFileName, finalString)
+  s.stop('Creating .env file')
   consola.success(`${ envFileName } file created successfully`)
 }
 
@@ -63,6 +66,30 @@ export async function getEnvVariables(projectId: string, environment: string): P
     return variables
   } catch (e) {
     cancel('Failed to fetch variables')
+    process.exit(0)
+  }
+}
+
+export async function pushEnvFile(variables: Env[], projectId: string, environment: string): Promise<void> {
+  const api = await useApi()
+
+  s.start('Pushing variables')
+  try {
+    const body: VariablesCreateInput = {
+      method: 'overwrite',
+      projectId,
+      environment,
+      variables: variables.map((variable) => ({
+        key: variable.key,
+        value: variable.value,
+        projectId,
+        environment
+      }))
+    }
+    await api(`/variable`, { method: 'POST', body })
+    s.stop('Pushing variables')
+  } catch (e) {
+    cancel('Failed to push variables')
     process.exit(0)
   }
 }
